- Molflow 2.6.75 - GLCombo updates displayed text on SetUserValue

+++ FacetCoordinates.cpp
@@ -311,7 +311,7 @@ void FacetCoordinates::RemoveRow(size_t rowId){
 			facetListC->SetSelectedRow((int)rowId);
 		}
 		else { //Select last
-			facetListC->SetSelectedRow(facetListC->GetNbRow()-1);
+			facetListC->SetSelectedRow((int)facetListC->GetNbRow()-1);
 		}
 		int selRow = facetListC->GetSelectedRow() + 1;
 		insertBeforeButton->SetEnabled(selRow);
